// Application/Frontend/frontend-app/src/app/generated/trade-client/model/transferDirection.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export const TransferDirection = {
    BankToWallet: 'BANK_TO_WALLET',
    WalletToBank: 'WALLET_TO_BANK',
} as const;
export type TransferDirection = typeof TransferDirection[keyof typeof TransferDirection];