// Application/Frontend/frontend-app/src/app/generated/trade-client/model/transferResponse.ts
/**
 * Trade REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransferDirection } from './transferDirection';


export interface TransferResponse { 
    transferId: string;
    accountId: number;
    direction: TransferDirection;
    amount: number;
    /**
     * The wallet (trading) balance after the transfer.
     */
    walletBalance: number;
    /**
     * The linked bank account\'s balance after the transfer.
     */
    bankBalance: number;
    createdAt: string;
}
export namespace TransferResponse {
}